+++ app/src/main/java/com/example/swith/data/Group.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Group(
     val announcementContent: String,
     val attendanceRate: Int,
-    val groupIdx: Int,
+    val groupIdx: Long,
     val groupImageUrl: String?,
     val interestContent: String,
     val memberLimit: Int,
