+++ app/src/main/java/com/example/swith/repository/home/HomeRepository.kt
@@ -9,6 +9,6 @@ import com.example.swith.utils.error.RemoteErrorEmitter
 import kotlinx.coroutines.flow.Flow
 
 class HomeRepository(private val homeRemoteDataSource: HomeRemoteDataSource){
-    suspend fun getAllStudy(remoteErrorEmitter: RemoteErrorEmitter, userId: Int) : GroupList? = homeRemoteDataSource.getAllStudy(remoteErrorEmitter, userId)
+    suspend fun getAllStudy(remoteErrorEmitter: RemoteErrorEmitter, userId: Long) : GroupList? = homeRemoteDataSource.getAllStudy(remoteErrorEmitter, userId)
 //    suspend fun getRegPlace(remoteErrorEmitter: RemoteErrorEmitter, regionIdx: Long) = homeRemoteDataSource.getRegPlace(remoteErrorEmitter, regionIdx)
 }
